feat: extract BecomeASponsor component and reorder sections for improved user flow
- Extract "Become a Sponsor" section from Partners component into separate BecomeASponsor component
- Reorder sections in App: move Tickets before Partners, CFP after Charity, and add BecomeASponsor at the end
- Update navigation and footer to reflect new section order (Tickets, Partners, Speakers, Venue, CFP)
- Remove "Get your ticket" CTA button from navigation
- Fix indentation in partners array (spaces to tabs)
