+++ src/components/navigation.tsx
@@ -19,10 +19,11 @@ export const Navigation = () => {
 	}, []);
 
 	const navItems = [
-		{ label: 'CFP', id: 'cfp' },
-		{ label: 'Sponsor', id: 'partners' },
+		{ label: 'Tickets', id: 'tickets' },
+		{ label: 'Partners', id: 'partners' },
 		{ label: 'Speakers', id: 'speakers' },
 		{ label: 'Venue', id: 'venue' },
+		{ label: 'CFP', id: 'cfp' },
 		{ label: 'Contact', id: 'contact' },
 	];
 
@@ -68,13 +69,6 @@ export const Navigation = () => {
 								{item.label}
 							</a>
 						))}
-						<a
-							href="#tickets"
-							onClick={() => handleNavClick('Get your ticket')}
-							className="rounded-lg bg-meetjs-green px-6 py-2.5 text-sm font-bold text-black transition-all hover:bg-meetjs-green/90 hover:shadow-lg hover:shadow-meetjs-green/20"
-						>
-							Get your ticket
-						</a>
 					</div>
 
 					<button
@@ -114,13 +108,6 @@ export const Navigation = () => {
 									{item.label}
 								</a>
 							))}
-							<a
-								href="#tickets"
-								onClick={() => handleNavClick('Get your ticket')}
-								className="mt-2 rounded-lg bg-meetjs-green px-6 py-3 text-center text-base font-bold text-black transition-all hover:bg-meetjs-green/90"
-							>
-								Get your ticket
-							</a>
 						</div>
 					</div>
 				)}
@@ -128,4 +115,3 @@ export const Navigation = () => {
 		</nav>
 	);
 };
-
